// src/API.ts
export type ModelStringInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};

export enum ModelAttributeTypes {
  binary = 'binary',
  binarySet = 'binarySet',
  bool = 'bool',
  list = 'list',
  map = 'map',
  number = 'number',
  numberSet = 'numberSet',
  string = 'string',
  stringSet = 'stringSet',
  _null = '_null',
}

export type ModelSizeInput = {
  ne?: number | null;
  eq?: number | null;
  le?: number | null;
  lt?: number | null;
  ge?: number | null;
  gt?: number | null;
  between?: Array<number | null> | null;
};

export type ModelBooleanInput = {
  ne?: boolean | null;
  eq?: boolean | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
};

export type Todo = {
  __typename: 'Todo';
  id: string;
  title: string;
  message: string;
  isDone: boolean;
  dueDate?: string | null;
  createdAt: string;
  type: string;
  updatedAt: string;
  owner?: string | null;
};

export type UpdateTodoInput = {
  id: string;
  title?: string | null;
  message?: string | null;
  isDone?: boolean | null;
  dueDate?: string | null;
  createdAt?: string | null;
  type?: string | null;
};

export type DeleteTodoInput = {
  id: string;
};

export type ModelTodoFilterInput = {
  id?: ModelIDInput | null;
  title?: ModelStringInput | null;
  message?: ModelStringInput | null;
  isDone?: ModelBooleanInput | null;
  dueDate?: ModelStringInput | null;
  createdAt?: ModelStringInput | null;
  type?: ModelStringInput | null;
  and?: Array<ModelTodoFilterInput | null> | null;
  or?: Array<ModelTodoFilterInput | null> | null;
  not?: ModelTodoFilterInput | null;
};

export type ModelIDInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};

export type ModelTodoConnection = {
  __typename: 'ModelTodoConnection';
  items: Array<Todo | null>;
  nextToken?: string | null;
};

export type ModelStringKeyConditionInput = {
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
};

export enum ModelSortDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}

export type CreateTodoMutationVariables = {
  input: CreateTodoInput;
  condition?: ModelTodoConditionInput | null;
};

export type CreateTodoMutation = {
  createTodo?: {
    __typename: 'Todo';
    id: string;
    title: string;
    message: string;
    isDone: boolean;
    dueDate?: string | null;
    createdAt: string;
    type: string;
    updatedAt: string;
    owner?: string | null;
  } | null;
};

export type UpdateTodoMutationVariables = {
  input: UpdateTodoInput;
  condition?: ModelTodoConditionInput | null;
};

export type UpdateTodoMutation = {
  updateTodo?: {
    __typename: 'Todo';
    id: string;
    title: string;
    message: string;
    isDone: boolean;
    dueDate?: string | null;
    createdAt: string;
    type: string;
    updatedAt: string;
    owner?: string | null;
  } | null;
};

export type DeleteTodoMutationVariables = {
  input: DeleteTodoInput;
  condition?: ModelTodoConditionInput | null;
};

export type DeleteTodoMutation = {
  deleteTodo?: {
    __typename: 'Todo';
    id: string;
    title: string;
    message: string;
    isDone: boolean;
    dueDate?: string | null;
    createdAt: string;
    type: string;
    updatedAt: string;
    owner?: string | null;
  } | null;
};

export type GetTodoQueryVariables = {
  id: string;
};

export type GetTodoQuery = {
  getTodo?: {
    __typename: 'Todo';
    id: string;
    title: string;
    message: string;
    isDone: boolean;
    dueDate?: string | null;
    createdAt: string;
    type: string;
    updatedAt: string;
    owner?: string | null;
  } | null;
};
// ...
export type GetTodosByDateQuery = {
  getTodosByDate?: {
    __typename: 'ModelTodoConnection';
    items: Array<{
      __typename: 'Todo';
      id: string;
      title: string;
      message: string;
      isDone: boolean;
      dueDate?: string | null;
      createdAt: string;
      type: string;
      updatedAt: string;
      owner?: string | null;
    } | null>;
    nextToken?: string | null;
  } | null;
};